requirements.txt
asgiref==3.7.2
celery==5.3.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.60.0
ruff==0.1.11
gunicorn==21.2.0
Pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
uvicorn==0.25.0